+++ src/Commands/LarrockManagerCommand.php
@@ -18,7 +18,7 @@ class LarrockManagerCommand extends Command
      *
      * @var string
      */
-    protected $description = 'Manage larrockCMS packages';
+    protected $description = 'Package manager for LarrockCMS';
 
     /**
      * Execute the console command.
@@ -27,34 +27,33 @@ class LarrockManagerCommand extends Command
      */
     public function handle()
     {
-        $this->info('=== LarrockCMS Manager ===');
+        $this->info('=== LarrockCMS Package manager ===');
 
         $packages = [
-            'Do not install other packages', 'larrock-catalog', 'larrock-cart', 'larrock-wizard', 'larrock-discount',
-            'larrock-feed', 'larrock-category', 'larrock-reviews', 'larrock-smartbanners',
-            'larrock-menu', 'larrock-users', 'larrock-pages', 'larrock-blocks', 'larrock-contact',
-            'larrock-admin-seo', 'larrock-search', 'larrock-yandex-kassa', 'all'
+            'fanamurov/larrock-catalog', 'fanamurov/larrock-cart', 'fanamurov/larrock-wizard',
+            'fanamurov/larrock-discount', 'fanamurov/larrock-feed', 'fanamurov/larrock-category',
+            'fanamurov/larrock-reviews', 'fanamurov/larrock-smartbanners', 'fanamurov/larrock-menu',
+            'fanamurov/larrock-users', 'fanamurov/larrock-pages', 'fanamurov/larrock-blocks',
+            'fanamurov/larrock-contact', 'fanamurov/larrock-admin-seo', 'fanamurov/larrock-search',
+            'fanamurov/larrock-yandex-kassa', 'fanamurov/larrock-vscale'
         ];
 
-        $name = $this->choice('What to install/update?', $packages);
+        $question = ['All', $packages, 'Do not install other packages'];
+
+        $name = $this->choice('What to install/update?', $question);
 
         if($name === 'all') {
             $this->info('Install all packages LarrockCMS');
-            foreach ($packages as $package) {
-                if ($package !== 'all' && $package !== 'Do not install other packages') {
-                    $this->info('composer require fanamurov/' . $package .':^1.0 --prefer-dist');
-                    echo shell_exec('composer require fanamurov/' . $package .':^1.0 --prefer-dist');
-                }
-            }
-        }elseif($name === 'Do not install other packages'){
-            $this->info('Do not install other packages');
-        }else{
+            echo shell_exec('composer require '. implode(':^1.0 ', $packages) .' --prefer-dist');
+        }elseif($name !== 'Do not install other packages'){
             $this->info('composer require fanamurov/'. $name .':^1.0 --prefer-dist');
             echo shell_exec('composer require fanamurov/'. $name .':^1.0 --prefer-dist');
         }
 
-        $this->call('vendor:publish');
-        $this->call('migrate');
+        if($name !== 'Do not install other packages'){
+            $this->call('vendor:publish');
+            $this->call('migrate');
+        }
 
         $this->info('The task is completed! Thank you for using LarrockCMS');
     }